feat: Add D-Bus interface for interactive use
Caterpillar is turned into a state machine driven application, that is
designed to be running in the background.

A D-Bus interface allows to interact with the application and to trigger
certain actions such as:
* searching for an update
* installing (and optionally rebooting) or skipping the installation of
  an update
These blocking actions, as well as the state handling are run in
separate tokio tasks (threads).
Various state related facilities are exposed as properties.
More information on the interface can be found in the XML describing it.

Further integration for D-Bus and systemd is added/updated to fully
integrate also with scenarios in which caterpillar is started over D-
Bus.

Signed-off-by: David Runge <dave@sleepmap.de>